Updated recipe to use ruby_install instead of ruby_build
I'm not using the LWRP from ruby_install recipe as I don't like
having to deal with chruby, so I built my own.

This is a better alternative as It doesn’t require chruby and
defaults to system-wide install, which you will thank me when
you try to run anything on cron.